+++ app/src/main/java/com/example/myfinance/ui/common/FinappTopBar.kt
@@ -1,0 +1,40 @@
+package com.example.myfinance.ui.common
+
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun FinappTopBar(title: String, actionIcon: Int? = null, action: () -> Unit = {}) {
+    CenterAlignedTopAppBar(
+        title = { Text(title, fontSize = 22.sp) },
+        colors = TopAppBarDefaults.topAppBarColors(
+            containerColor = MaterialTheme.colorScheme.primary,
+            titleContentColor = MaterialTheme.colorScheme.onPrimary
+        ),
+        actions = {
+            actionIcon?.let {
+                IconButton(onClick = action) {
+                    Icon(
+                        ImageVector.vectorResource(actionIcon),
+                        contentDescription = "История",
+                        modifier = Modifier.size(24.dp),
+                        tint = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+        }
+    )
+}
